# Swagger/api_doc.yaml
swagger: '2.0'
info:
  title: WEConnect API
  description: The API for all the WEConnect operations.
  termsOfService: ""
  contact:
    email: "amaechi@gmail.com"
  license:
    name: "Apache 2.0"
  version: version 1.0
host: '127.0.0.1:4000'
schemes:
  - http
basePath: /
produces:
  - application/json
consumes:
  - application/json
paths:
  /businesses:
    get:
      summary: Get all businesses
      responses:
        '200':
          description: An array of businesses
          schema:
            $ref: '#/definitions/BusinessObjects'
        '401':
          description: Auth failed
        '404':
          description: No businesses found
        '500':
          description: General server error
    post:
      summary: "Register a business"
      description: ""
      operationId: "add a business"
      consumes:
      - "application/json"
      produces:
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Business object that needs to be added to the Businesses"
        required: true
        schema:
          $ref: "#/definitions/BusinessObjects"
      responses:
        405:
          description: "Invalid input"
  /businesses-category:
    get:
      summary: "Find A business by category"
      description: "Returns an array of Business by category"
      operationId: ""
      produces:
      - "application/json"
      parameters:
      - name: "location"
        in: "query"
        description: "category of businesses to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BusinessObjects"
        400:
          description: "Invalid location supplied"
        404:
          description: "businesses not found"
        '500':
          description: General server error
  /businesses-location:
    get:
      summary: "Find A business by location"
      description: "Returns an array of Business"
      operationId: ""
      produces:
      - "application/json"
      parameters:
      - name: "location"
        in: "query"
        description: "ID of businesses to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BusinessObjects"
        400:
          description: "Invalid location supplied"
        404:
          description: "businesses not found"
        '500':
          description: General server error
  /businesses/{businessid}:
    get:
      summary: "Get a business"
      description: "Returns an array of Business"
      operationId: ""
      produces:
      - "application/json"
      parameters:
      - name: "businessid"
        in: "path"
        description: "ID of businesses to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BusinessObjects"
        400:
          description: "Invalid location supplied"
        404:
          description: "businesses not found"
        '500':
          description: General server error
    put:
      summary: "Update business profile"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "businessid"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
    delete:
      summary: "Remove a business"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "businessid"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
  /businesses/{businessid}/reviews:
    get:
      summary: "Get all reviews for a business"
      description: "Returns all reviews for a business"
      produces:
      - "application/json"
      parameters:
      - name: "businessid"
        in: "path"
        description: "ID of business to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/BusinessObjects"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Business not found"
    post:
      summary: "Add a review for a business"
      description: ""
      operationId: "updateBusinessWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "businessid"
        in: "path"
        description: "businessID of Business that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "status"
        in: "formData"
        description: "Updated status of the Business"
        required: false
        type: "string"
      responses:
        405:
          description: "Invalid input"
  /auth/signup:
    post:
      summary: "Register a user"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "List of user object"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
  /auth/login:
    post:
      summary: "Login a user"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "List of login object"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
definitions: #schema definitions below
  BusinessObjects:
    type: object
    properties:
      Businesses:
        type: array
        items:
          type: string
  BusinessObject:
    type: object
    properties:
      DNIS:
        type: string
        description: some description
      OfficeKey:
        type: string
      Action:
        type: string
      NextActionWH:
        type: string
      ActionType:
        type: string
      NextActionOWH:
        type: string
      CallType:
        type: string
      Dequeue:
        type: boolean
      WaitTime:
        type: integer
        format: int32
      SkipIfNoReady:
        type: boolean
        